Bug Fix: clener will crush on wrong file name

+++ internal/cleaning/cleaning.go
@@ -22,10 +22,12 @@ func cleaning(currenttime time.Time, period int) {
 			for _, subitem := range subitems {
 				if !subitem.IsDir() {
 					// fmt.Println(subitem.Name())
-					tm, _ := time.Parse("2006-01-02_15:04:05", subitem.Name()[:19])
-					if currenttime.Sub(tm).Hours() > float64(period) {
-						os.Remove("/home/media/video/" + item.Name() + "/" + subitem.Name())
-						fmt.Println("Removed " + item.Name() + "/" + subitem.Name())
+					if len(subitem.Name()) > 19 {
+						tm, err := time.Parse("2006-01-02_15:04:05", subitem.Name()[:19])
+						if currenttime.Sub(tm).Hours() > float64(period) && err == nil {
+							os.Remove("/home/media/video/" + item.Name() + "/" + subitem.Name())
+							fmt.Println("Removed " + item.Name() + "/" + subitem.Name())
+						}
 					}
 				}
 			}
